Implement #83: Attendee names
This adds another XML-table to the application which consists as a static list
of attendees.

Each appearence of attendees as item owner or buyer has been replaced/enriched
for displaying purposes with "<nickname> (#<regid>)" or for input fields,
HTML5's "list" attribute is used like it already is in other parts of the
application.

A downside of this is that browsers currently show the "value" which is still
just the registration ID in the input field instead of the "label". But when
searching the field (that is, when clicking on it and/or typing in it), an
automatically filtered (by current input) list of attendee names is shown.

IMO, as a first step, using this browser feature (tested with Firefox and
Chrome) without additional Javascript requirements is a very good trade off.
Let's see if someone complains about the input fields only containing numbers
and we can change it still later on.