Change artshow recipie
This implements ticket #84.

It's formatting the recipie in monospace with the required printer's column
width. IMO this makes it easier to develop and test stuff because one cannot
know what the browser/printer will do with normal HTML and word wrapping until
one printed.

The czech template uses the english template.

It's using a 3rd party module `texttable` for formatting the tables. I included
it raw via file because the mechanism for installing python packages seems very
complicated in this application. It's only one file after all.